http: unify the buffer limit to uint64 (#42608)
Commit Message:http: unify the buffer limit to uint64
Additional Description:
As the name.
Risk Level: low.
Testing: n/a.
Docs Changes: n/a.
Release Notes: n/a.

+++ source/common/http/filter_manager.cc
@@ -1896,11 +1896,11 @@ void ActiveStreamEncoderFilter::onEncoderFilterBelowWriteBufferLowWatermark() {
   parent_.callLowWatermarkCallbacks();
 }
 
-void ActiveStreamEncoderFilter::setEncoderBufferLimit(uint32_t limit) {
+void ActiveStreamEncoderFilter::setEncoderBufferLimit(uint64_t limit) {
   parent_.setBufferLimit(limit);
 }
 
-uint32_t ActiveStreamEncoderFilter::encoderBufferLimit() { return parent_.buffer_limit_; }
+uint64_t ActiveStreamEncoderFilter::encoderBufferLimit() { return parent_.buffer_limit_; }
 
 void ActiveStreamEncoderFilter::continueEncoding() { commonContinue(); }
